Move tx serialization to transaction helpers
